ProductCategory Application services has added to the solution.
Includes:
On branch dev
Changes to be committed:
	modified:   TopraqShop.sln
	modified:   TopraqShop.sln.DotSettings
	new file:   src/Framework/TopraqShop.Framework.Base/Application/GenerateSlug.cs
	new file:   src/Framework/TopraqShop.Framework.Base/Application/OperationResult.cs
	modified:   src/Framework/TopraqShop.Framework.Base/TopraqShop.Framework.Base.csproj
	new file:   src/ShopManagement/Domain/Application/TopraqShop.ShopManagement.Domain.Application.Base/ProductCategory/ProductCategoryApplication.cs
	new file:   src/ShopManagement/Domain/Application/TopraqShop.ShopManagement.Domain.Application.Base/TopraqShop.ShopManagement.Domain.Application.Base.csproj
	renamed:    src/ShopManagement/Domain/Application/TopraqShop.ShopManagement.Domain.Application.Contracts/ProductCategory/CreateProductCategory.cs -> src/ShopManagement/Domain/Application/TopraqShop.ShopManagement.Domain.Application.Contracts/ProductCategory/Concrete/CreateProductCategory.cs
	renamed:    src/ShopManagement/Domain/Application/TopraqShop.ShopManagement.Domain.Application.Contracts/ProductCategory/EditProductCategory.cs -> src/ShopManagement/Domain/Application/TopraqShop.ShopManagement.Domain.Application.Contracts/ProductCategory/Concrete/EditProductCategory.cs
	renamed:    src/ShopManagement/Domain/Application/TopraqShop.ShopManagement.Domain.Application.Contracts/ProductCategory/ProductCategorySearchModel.cs -> src/ShopManagement/Domain/Application/TopraqShop.ShopManagement.Domain.Application.Contracts/ProductCategory/Concrete/ProductCategorySearchModel.cs
	renamed:    src/ShopManagement/Domain/Application/TopraqShop.ShopManagement.Domain.Application.Contracts/ProductCategory/ProductCategoryViewModel.cs -> src/ShopManagement/Domain/Application/TopraqShop.ShopManagement.Domain.Application.Contracts/ProductCategory/Concrete/ProductCategoryViewModel.cs
	deleted:    src/ShopManagement/Domain/Application/TopraqShop.ShopManagement.Domain.Application.Contracts/ProductCategory/IProductCategoryApplication.cs
	new file:   src/ShopManagement/Domain/Application/TopraqShop.ShopManagement.Domain.Application.Contracts/ProductCategory/Interface/IProductCategoryApplication.cs
	modified:   src/ShopManagement/Domain/TopraqShop.ShopManagement.Domain.Base/ProductCategoryAggregate/IProductCategoryRepository.cs
	modified:   src/ShopManagement/Domain/TopraqShop.ShopManagement.Domain.Base/ProductCategoryAggregate/ProductCategoryBase.cs
	new file:   test/Framework/TopraqShop.Framework.Base.Tests.Unit/Application/GenerateSlugTests.cs
	new file:   test/Framework/TopraqShop.Framework.Base.Tests.Unit/Application/OperationResultTests.cs
	new file:   test/Framework/TopraqShop.Framework.Base.Tests.Unit/TopraqShop.Framework.Base.Tests.Unit.csproj
	new file:   test/ShopManagement/Application/TopraqShop.ShopManagement.Domain.Application.Base.Tests.Unit/ProductCategory/ProductCategoryAppTests.cs
	new file:   test/ShopManagement/Application/TopraqShop.ShopManagement.Domain.Application.Base.Tests.Unit/TopraqShop.ShopManagement.Application.Base.Tests.Unit.csproj

+++ src/ShopManagement/Domain/TopraqShop.ShopManagement.Domain.Base/ProductCategoryAggregate/ProductCategoryBase.cs
@@ -3,7 +3,7 @@ using TopraqShop.Framework.Base.Domain;
 
 namespace TopraqShop.ShopManagement.Domain.Base.ProductCategoryAggregate
 {
-    public class ProductCategoryBase : EntityBase<int>
+    public class ProductCategoryBase : EntityBase<byte>
     {
         public string Name { get; private set; }
         public string Description { get; private set; }
